refactor: split LogLayer.basic.test.ts into smaller files
Extract errorOnly (62 lines), integration tests (213 lines),
metadata/error/context flow (187 lines), and raw method (908 lines)
into separate test files. Reduces basic test file from 1622 lines (48KB)
to 316 lines (8.6KB) to avoid biome linter OOM.
